Remove EDH fitness activities api methods & tests
